fixed bugs and reduced the cyclomatic complexity
